Clarified usage of exit.
Exiting on an ISE closes the UI, which is bad.
Additionally exit in a sub script (e.g. called via & script.ps1) may swallow all output. This occurs primarily when the result of a script is stored in a variable and there was an error. The error will never be reported then.